requirements.txt
openai
faiss-cpu
sentence-transformers
pypdf
tiktoken
fastapi
uvicorn
python-dotenv
pyyaml
langchain-chroma
langchain==0.1.20
chromadb==0.4.24
sentence-transformers
python-dotenv